pyproject.toml
[project]
name = "domain-search"
version = "0.1.0"
description = "CLI tool to search for available domain names across all known TLDs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "rich>=13.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
